Split Supabase client into server and browser modules
