+++ talentify-next-frontend/lib/supabase/client.ts
@@ -1,0 +1,10 @@
+import { createBrowserClient } from '@supabase/ssr'
+import type { Database } from '@/types/supabase'
+
+export const createClient = () =>
+  createBrowserClient<Database>(
+    process.env.NEXT_PUBLIC_SUPABASE_URL!,
+    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
+  )
+
+export type { Database }
