+++ talentify-next-frontend/lib/supabase/index.ts
@@ -1,32 +1,3 @@
-import { createBrowserClient, createServerClient, type CookieOptions } from '@supabase/ssr'
-import { cookies } from 'next/headers'
-import type { Database } from '@/types/supabase'
-
-export const createBrowserSupabaseClient = () =>
-  createBrowserClient<Database>(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
-  )
-
-export async function createServerSupabaseClient() {
-  const cookieStore = await cookies()
-  return createServerClient<Database>(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-    {
-      cookies: {
-        get(name: string) {
-          return cookieStore.get(name)?.value
-        },
-        set(name: string, value: string, options: CookieOptions) {
-          cookieStore.set({ name, value, ...options })
-        },
-        remove(name: string, options: CookieOptions) {
-          cookieStore.set({ name, value: '', ...options })
-        }
-      }
-    }
-  )
-}
-
-export type { Database }
+export { createClient as createBrowserSupabaseClient } from './client'
+export { createClient as createServerSupabaseClient, runtime } from './server'
+export type { Database } from './types'
